remove core_types dependency on lib3h_protocol
