+++ core_types/Cargo.toml
@@ -41,7 +41,6 @@ regex = "=1.1.2"
 shrinkwraprs = "=0.2.1"
 crossbeam-channel = "=0.3.8"
 lib3h_crypto_api = "=0.0.16"
-lib3h_protocol = "=0.0.16"
 parking_lot ="=0.9.0"
 log = "=0.4.8"
 holochain_logging = { version="=0.0.1", path = "../logging" }
